mcmc.GP

Function to sample from a GP model that is assumed to be a -lnLikelihood surface with flat priors and bounded region

mcmc.GP(
  gp,
  x0,
  nmcmc,
  rng,
  format = "lin",
  mix = "joint",
  splinefuns = NULL,
  jmp0 = 0.35 * (rng[, 2] - rng[, 1]),
  ar.target = 0.5,
  priors = NA,
  settings,
  run.block = TRUE,
  n.of.obs,
  llik.fn,
  hyper.pars,
  resume.list = NULL
)

Arguments

gp

Gaussian Process

x0

initial values

nmcmc

number of iterations

rng

range of knots

format

lin = lnlike fcn, log = log(lnlike)

mix

each = jump each dim. independently, joint = jump all at once

splinefuns

spline functions, not used

jmp0

initial jump variances

ar.target

acceptance rate target

priors

prior list

settings

PEcAn settings list

run.block

is this a new run or making the previous chain longer

n.of.obs

number of observations

llik.fn

list that contains likelihood functions

hyper.pars

hyper parameters

resume.list

list of needed info if we are running the chain longer

Author

Michael Dietze