Helper function to sample from previous MCMC chain while proposing new knots

Helper function to sample from previous MCMC chain while proposing new knots

sample_MCMC(
  mcmc_path,
  n.param.orig,
  prior.ind.orig,
  n.post.knots,
  knots.params.temp,
  prior.list,
  prior.fn,
  sf,
  sf.samp
)

Arguments

mcmc_path

path to previous emulator mcmc samples object

n.param.orig

vector, number of parameters targeted in each (pft) sublist

prior.ind.orig

list, actual indices of parameters targeted in each (pft) sublist

n.post.knots

number of new samples requested

knots.params.temp

list of parameter samples proposed from the original PDA-prior

prior.list

PDA-prior list

prior.fn

list for parameter d/r/q/p functions

sf

SF parameter names

sf.samp

SF parameters MCMC samples

Author

Istem Fer